Progressed connection, need to fix Dolphin loop jank and game_watcher debugging

+++ worlds/custom_robo/helpers.py
@@ -23,7 +23,7 @@ class CRRamData(NamedTuple):
     in_game_room_id: Optional[int] = None
     item_count: Optional[int] = None
 
-async def wait_for_next_loop(time_to_wait: int):
+async def wait_for_next_loop(time_to_wait: float):
     await asyncio.sleep(time_to_wait)
 
 class StringByteFunction:
